catalog-service/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
mysqlclient==2.2.0
requests==2.31.0
